requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
python-socketio==5.9.0
python-engineio==4.7.1
requests==2.31.0
python-dotenv==1.0.0
websocket-client==1.6.4
rel==0.4.9
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
scikit-learn==1.3.0
pandas==2.0.3
